Use unsigned immediates for arithmetic functions
Given the use of 2's complement integer representation, the machine code
is the same either way. The use of signed immediates was done to
simplify the implementation of the s390x backend, and for consistency
with the C version, but resulted in a more fragile internal API, so
instead, the s390x version will need to deal with messy signed/unsigned
casts.

Reverts 9f5adef734e55de62f35bd758af19a224b9a4358