Rework arm64 backend based on riscv backend
Now uses a RawReg type for some internal operations, and hard-codes a
specific scratch register of that type (x17) for use as a temporary
register, replacing the aux_reg function

This opened the road to significant streamlining - for instance, it
became clear that store_to_byte and load_from_byte only ever used
TEMP_REG as a second parameter, so it could be made into a
single-parameter function.

The scratch registers were elimiated from the Arm64Register enum, and
tests were adjusted to account for that.